Fix tensor formating in the notebook.
* Add commas after each row for 2D tensor formating
* Adjust notebook.html (remove lets)
