+++ tools/test.sh
@@ -11,6 +11,7 @@ node dist/backend_test.js
 node dist/util_test.js
 node dist/mnist_test.js
 node dist/nb_transpiler_test.js
+node dist/format_test.js
 
 # Now using the Deep Learn backend.
 export PROPEL="dl"
@@ -19,3 +20,4 @@ node dist/backend_test.js
 node dist/util_test.js
 node dist/mnist_test.js
 node dist/nb_transpiler_test.js
+node dist/format_test.js
