+++ website/notebook.html
@@ -15,15 +15,18 @@
     <div id=cells>
 <script type="notebook">
 import { $ } from "propel";
-console.log("5*5 =", $(5).mul(5));
+
+t = $([[2, 3], [30, 20]])
+
+t.mul(5)
 </script>
 
 <script type="notebook">
 import { grad, linspace } from "propel";
 import { plot } from "matplotlib";
 
-let f = (x) => $(x).mul(x);
-let x = linspace(-4, 4, 200);
+f = (x) => $(x).mul(x);
+x = linspace(-4, 4, 200);
 plot(x, f(x),
      x, grad(f)(x));
 </script>
@@ -39,8 +42,8 @@ plot(x, tanh(x),
 </script>
 
 <script type="notebook">
-let f = await fetch('mnist/README');
-let t = await f.text();
+f = await fetch('mnist/README');
+t = await f.text();
 t;
 </script>
     </div> <!-- #cells -->
